Remove local user admin customization from resources to pages
